requirements.txt
requests
psycopg2-binary
python-dotenv
clickhouse-connect
dbt-core
dbt-clickhouse